requirements.txt
requests
beautifulsoup4
playwright
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
pandas
python-dotenv
google-genai
